chore: relocate test so each one tests its own stuff. Improve docs in code.
