+++ regimes/nz/identities.go
@@ -6,6 +6,7 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/i18n"
 	"github.com/invopop/gobl/org"
+	"github.com/invopop/gobl/pkg/here"
 	"github.com/invopop/validation"
 )
 
@@ -17,6 +18,23 @@ var identityKeyDefinitions = []*cbc.Definition{
 		Key: IdentityKeyIRD,
 		Name: i18n.String{
 			i18n.EN: "IRD Number",
+		},
+		Desc: i18n.String{
+			i18n.EN: here.Doc(`
+				The IRD number is the primary tax identifier in New Zealand, issued by
+				Inland Revenue to individuals, companies, trusts, and other entities.
+				Format: 8 or 9 digits (XXX-XXX-XXX when displayed). Valid range:
+				10,000,000 to 200,000,000. For GST-registered businesses, the IRD
+				number also serves as their GST number.
+			`),
+		},
+		Sources: []*cbc.Source{
+			{
+				Title: i18n.String{
+					i18n.EN: "IRD Numbers Overview",
+				},
+				URL: "https://www.ird.govt.nz/managing-my-tax/ird-numbers",
+			},
 		},
 	},
 }
