+++ regimes/nz/org_identities.go
@@ -7,6 +7,7 @@ import (
 	"github.com/invopop/gobl/i18n"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/pkg/gs1"
+	"github.com/invopop/gobl/pkg/here"
 	"github.com/invopop/validation"
 )
 
@@ -19,7 +20,27 @@ var orgIdentityDefinitions = []*cbc.Definition{
 			i18n.EN: "NZ Business Number",
 		},
 		Desc: i18n.String{
-			i18n.EN: "13-digit identifier based on the GS1 Global Location Number (GLN) standard, starting with NZ prefix 94.",
+			i18n.EN: here.Doc(`
+				The NZBN is a globally unique 13-digit identifier automatically assigned to
+				companies registered with the New Zealand Companies Office. Other entities
+				can apply voluntarily. Based on the GS1 Global Location Number (GLN) standard,
+				it always starts with the New Zealand GS1 prefix '94'. For Peppol e-invoicing,
+				the NZBN is used with the '0088:' scheme identifier (e.g., 0088:9429041234567).
+			`),
+		},
+		Sources: []*cbc.Source{
+			{
+				Title: i18n.String{
+					i18n.EN: "NZBN Official",
+				},
+				URL: "https://www.nzbn.govt.nz/whats-an-nzbn/about/",
+			},
+			{
+				Title: i18n.String{
+					i18n.EN: "GS1 Check Digit Calculator",
+				},
+				URL: "https://www.gs1.org/services/check-digit-calculator",
+			},
 		},
 	},
 }
